fix(core): use stable key-order-independent stringify in CacheKeyGenerator.fromDBParts
PostgreSQL jsonb columns (mastra_workflow_snapshot) normalize object key
order while text columns (mastra_messages) preserve insertion order. This
caused CacheKeyGenerator.fromDBParts to produce different cache keys for
functionally-equal data-* parts, making messagesAreEqual return false and
minting duplicate message copies with the same rs_* reasoning id — which
OpenAI rejects with 'AI_APICallError: Duplicate item found'.

- Extract stableStringify (key-order-independent JSON.stringify) into a
  shared util at agent/message-list/cache/stable-stringify.ts
- Use stableStringify in CacheKeyGenerator.fromDBParts instead of
  JSON.stringify for data-* parts
- Remove duplicate stableStringify from response-cache.ts and import
  the shared version instead

Fixes #17432
