+++ packages/core/src/agent/message-list/cache/CacheKeyGenerator.ts
@@ -1,4 +1,5 @@
 import type { UIMessage as UIMessageV4 } from '@internal/ai-sdk-v4';
+import { stableStringify } from './stable-stringify';
 import * as AIV5 from '@internal/ai-sdk-v5';
 
 import { getImageCacheKey } from '../prompt/image-utils';
@@ -109,7 +110,7 @@ export class CacheKeyGenerator {
       if (part.type.startsWith('data-')) {
         // Stringify data for proper cache key comparison since data can be any type
         const data = (part as AIV5Type.DataUIPart<AIV5.UIDataTypes>).data;
-        key += JSON.stringify(data);
+        key += stableStringify(data); // order-independent: jsonb vs text storage may reorder keys
       } else {
         // Cast to UIMessageV4Part since we've already handled data-* parts above
         key += CacheKeyGenerator.fromAIV4Part(part as UIMessageV4Part);
