+++ packages/core/src/agent/message-list/cache/index.ts
@@ -1,2 +1,3 @@
 export { CacheKeyGenerator } from './CacheKeyGenerator';
 export type { MastraMessageContentV2, MastraMessagePart, UIMessageV4Part } from './types';
+export { stableStringify } from './stable-stringify';
